Refactor item and permission access logic; enhance exception handling and validations
- Replaced `AuthService` with `UserProvider` for user management.
- Introduced `ItemAccessService` to centralize and streamline access checks.
- Improved validation for DTOs with `@Valid` annotations and added detailed messages.
- Replaced `OldVersionException` with `WrongVersionException` for clarity.
- Updated `GlobalExceptionHandler` to support new exception types and validation errors.
- Enhanced `JwtFilter` and `RateLimitFilter` responses with detailed error JSON.

+++ src/main/java/com/betacom/anynoteapi/item/dto/UpdateItemRequest.java
@@ -8,5 +8,5 @@ import java.io.Serializable;
 /**
  * DTO for {@link Item}
  */
-public record UpdateItemRequest(String title, String content, @NotNull Long version) implements Serializable {
+public record UpdateItemRequest(String title, String content, @NotNull Integer version) implements Serializable {
 }
